requirements/base.txt
Django
Pillow
coverage
django-auth-ldap-ng
django-oauth-toolkit==0.9.0
djangorestframework
gevent
gunicorn
psycopg2
python-memcached
django-simple-history
isort